moved vectorAlgebraFunctional.py to utility
